+++ Sources/Internal/Render/RHI/Common/rhi_ShaderSource.cpp
@@ -122,7 +122,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
     
     // parse properties/samplers
     
-    DynamicMemoryFile*  in = DynamicMemoryFile::Create( (const uint8*)src.c_str(), src.length()+1, DAVA::File::READ );
+    DynamicMemoryFile*  in = DynamicMemoryFile::Create( (const uint8*)src.c_str(), (uint32)(src.length()+1), DAVA::File::READ );
 
     if( in )
     {
@@ -285,7 +285,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
 
                 if( ss )
                 {
-                    int n = ss - tags.c_str();
+                    size_t n = ss - tags.c_str();
 
                     strncpy( storage, tags.c_str(), n );
                     storage[n] = '\0';
@@ -344,7 +344,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                     cbuf->regCount  = 0;
                 }
 
-                p.bufferindex = cbuf - &(buf[0]);
+                p.bufferindex = static_cast<uint32>(cbuf - &(buf[0]));
                 
                 if( p.type == ShaderProp::TYPE_FLOAT1  ||  p.type == ShaderProp::TYPE_FLOAT2  ||  p.type == ShaderProp::TYPE_FLOAT3 )
                 {
@@ -407,8 +407,8 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  fsampler2d_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
-                int         sn      = sname.length();
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      sn      = sname.length();
 
                 DVASSERT(sampler.size()<10);
                 char ch = line[mbegin+1];
@@ -436,8 +436,8 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  samplercube_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
-                int         sn      = sname.length();
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      sn      = sname.length();
 
                 DVASSERT(sampler.size()<10);
                 char ch = line[mbegin+1];
@@ -465,7 +465,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  ftexture2d_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
                 FastName    suid    ( sname );
                 
                 for( unsigned s=0; s!=sampler.size(); ++s )
@@ -473,7 +473,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                     if( sampler[s].uid == suid )
                     {
                         int sl = sprintf( line+mbegin, "%u", s );
-                        int sn = sname.length();
+                        size_t sn = sname.length();
                         DVASSERT(sn>=sl);
                         line[mbegin+sl] = ',';
                         if( sn > sl )
@@ -498,8 +498,8 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  vsampler2d_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
-                int         sn      = sname.length();
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      sn      = sname.length();
 
                 DVASSERT(sampler.size()<10);
                 char ch = line[mbegin+1];
@@ -527,7 +527,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  vtexture2d_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
                 FastName    suid    ( sname );
                 
                 for( unsigned s=0; s!=sampler.size(); ++s )
@@ -535,7 +535,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                     if( sampler[s].uid == suid )
                     {
                         int sl = sprintf( line+mbegin, "%u", s );
-                        int sn = sname.length();
+                        size_t sn = sname.length();
                         DVASSERT(sn>=sl);
                         line[mbegin+sl] = ',';
                         if( sn > sl )
@@ -560,7 +560,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                 #else
                 std::string sname;  texturecube_re.get_pattern( 1, &sname );
                 #endif
-                int         mbegin  = strstr( line, sname.c_str() ) - line;
+                size_t      mbegin  = strstr( line, sname.c_str() ) - line;
                 FastName    suid    ( sname );
                 
                 for( unsigned s=0; s!=sampler.size(); ++s )
@@ -568,7 +568,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
                     if( sampler[s].uid == suid )
                     {
                         int sl = sprintf( line+mbegin, "%u", s );
-                        int sn = sname.length();
+                        size_t sn = sname.length();
                         DVASSERT(sn>=sl);
                         line[mbegin+sl] = ',';
                         if( sn > sl )
@@ -742,7 +742,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
     {
         // check if any const-buffer has more than one bigarray-prop
         
-        for( unsigned b=0,b_end=buf.size(); b!=b_end; ++b )
+        for( size_t b=0,b_end=buf.size(); b!=b_end; ++b )
         {
             unsigned    bigarray_cnt = 0;
 
@@ -867,7 +867,7 @@ ShaderSource::Construct( ProgType progType, const char* srcText, const std::vect
             var_len += Snprintf(var_def + var_len, sizeof(var_def) - var_len, "    //--------\n");
             var_len += Snprintf(var_def + var_len, sizeof(var_def) - var_len, "\n\n");
             
-            unsigned var_pos = (prog-code.c_str()) + strlen("XPROG_BEGIN") + buf_len +2;
+            size_t var_pos = (prog-code.c_str()) + strlen("XPROG_BEGIN") + buf_len +2;
 
             code.insert( prog-code.c_str(), buf_def, buf_len );
             code.insert( var_pos, var_def, var_len );
@@ -1013,7 +1013,7 @@ ShaderSource::Save( DAVA::File* out ) const
 
     vdecl.Save( out );
 
-    WriteUI4( out, prop.size() );
+    WriteUI4( out, static_cast<uint32>( prop.size() ) );
     for( unsigned p=0; p!=prop.size(); ++p )
     {
         WriteS0( out, prop[p].uid.c_str() );
@@ -1028,7 +1028,7 @@ ShaderSource::Save( DAVA::File* out ) const
         out->Write( prop[p].defaultValue, 16*sizeof(float) );
     }
 
-    WriteUI4( out, buf.size() );
+    WriteUI4( out, static_cast<uint32>( buf.size() ) );
     for( unsigned b=0; b!=buf.size(); ++b )
     {
         WriteUI4( out, buf[b].storage );
@@ -1036,7 +1036,7 @@ ShaderSource::Save( DAVA::File* out ) const
         WriteUI4( out, buf[b].regCount );
     }
 
-    WriteUI4( out, sampler.size() );
+    WriteUI4( out, static_cast<uint32>( sampler.size() ) );
     for( unsigned s=0; s!=sampler.size(); ++s )
     {
         WriteUI4( out, sampler[s].type );
@@ -1101,7 +1101,7 @@ ShaderSource::ShaderVertexLayout() const
 uint32
 ShaderSource::ConstBufferCount() const
 {
-    return buf.size();
+    return static_cast<uint32>(buf.size());
 }
 
 
@@ -1164,7 +1164,7 @@ ShaderSource::_Reset()
 //------------------------------------------------------------------------------
 
 void
-ShaderSource::_AppendLine( const char* line, uint32 lineLen )
+ShaderSource::_AppendLine( const char* line, size_t lineLen )
 {
 code.append( line, lineLen );
 code.push_back( '\n' );
@@ -1372,7 +1372,7 @@ ShaderSourceCache::Save( const char* fileName )
     {
         WriteUI4( file, FormatVersion );
         
-        WriteUI4( file, Entry.size() );
+        WriteUI4( file, static_cast<uint32>(Entry.size()) );
 Logger::Info( "saving cached-shaders (%u) :", Entry.size() );
         for( std::vector<entry_t>::const_iterator e=Entry.begin(),e_end=Entry.end(); e!=e_end; ++e )
         {
